package pattern: Reversible pattern substitutions.
A *pattern.P represents a template string containing a number of pattern words,
which are named locations where substitution may occur. A pattern may be
"matched" against a string to produce a set of bindings of names to substrings;
or it may be "applied" to a set of bindings to produce a transformed string.