transform: Rework reversibility again.
Remove the R type, it was not worth the extra cognitive overhead.  Make all *T
values reversible, and equip them with a method to check whether the reverse is
information-preserving.

This reduces the amount of code in the package quite a bit, and doesn't make it
any harder to understand.