Add a silly example program: madlib
Remember Mad Libs? Probably not. You are probably young enough that the idea of
people sitting together in a place, filling out a sheet of paper with blanks
for missing words and then reading out the resulting derangement of the text
that results frmo their substitution is at best a quaint archaeism.

Nevertheless, there was once a time when people considered this activity to be
at least mildly amusing, at least some people, and assuming they were either
juvenile and bored or sufficiently intoxicated, or both.